+++ package/src/index.ts
@@ -4,4 +4,3 @@ export * from './server';
 export * from './client';
 export * from './theme_switcher';
 export * from './types';
-export * from './db';